backend/requirements.txt
# requirements.txt

# Veri İşleme
pandas
kagglehub

# RAG Mimarisi ve Framework
langchain
langchain-core
langchain-text-splitters
langchain-community
sentence-transformers
langchain-huggingface
langchain-chroma

# Google Gemini Modelleri (LLM ve Embedding)
langchain-google-genai
google-generativeai

# Vektör Veritabanı
chromadb

# Web Framework (Backend API için)
fastapi
uvicorn[standard]

# CORS için
python-multipart

# Gizli Bilgileri Yönetme
python-dotenv